deployment/requirements.txt
streamlit
pandas
numpy==1.26.4
scikit-learn==1.3.0
Pillow
plotly